more specs to Location for us available troops
The results are now matching the Playbook for Example 4,
for availability while maintaining COIN control.

What needs to happen next is determine the operational
needs of each space with excess US. This can probably be done
passing in a callback defining the operational needs. Starting
to feel a bit clunky now.