excess - find Locations with excess us
This turned out to be pretty simple to implement as
all of the necessary supporting methods were already
implemented. Along with finding Locations where excess
US Troops are located, the following items are part of
this commit:

1. rubocop ignores everything in spec for Block and Method
lengths.

2. Pending specs for previously implemented helper methods
are now implemented. Some of these methods may go to private
and the specs deleted.

3. Changed from `create` to `build` for factory calls. No need
to persist data for those particular specs.